Switch to using base classes.
EF7 seems to complain about inherited IDs, but I can't mark the base as unused since it is part of the framework.  I'm not sure what else to do despite a lot of searching - the original code seems to be the idiomatic way of doing things, but doesn't work correctly!